Split out steppers for Array, IndexedSeq, LinearSeq, Range, Vector
